# static/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "API de prueba",
    "description": "Una API para fines de prueba",
    "version": "1.0"
  },
  "paths": {
    "/login": {
      "post": {
        "summary": "Iniciar sesión",
        "description": "Permite a los usuarios iniciar sesión en la aplicación",
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "Credenciales del usuario",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "nombre": {
                  "type": "string",
                  "description": "Nombre del usuario"
                },
                "contraseña": {
                  "type": "string",
                  "description": "Contraseña del usuario"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Inicio de sesión exitoso"
          },
          "401": {
            "description": "Credenciales inválidas"
          }
        }
      }
    },
    "/register": {
      "post": {
        "summary": "Registrar usuario",
        "description": "Permite a los usuarios registrarse en la aplicación",
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "Información del usuario",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "nombre": {
                  "type": "string",
                  "description": "Nombre del usuario"
                },
                "contraseña": {
                  "type": "string",
                  "description": "Contraseña del usuario"
                },
                "dni": {
                  "type": "string",
                  "description": "DNI del usuario"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Registro exitoso"
          },
          "400": {
            "description": "Solicitud inválida"
          }
        }
      }
    },
    "/test": {
      "get": {
        "summary": "Prueba de autenticación",
        "description": "Permite probar la autenticación utilizando un JWT",
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "description": "JWT de autenticación",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Autenticación exitosa"
          },
          "401": {
            "description": "Autenticación fallida"
          }
        },
        "security": [
          {
            "JWT": []
          }
        ]
      }
    }
  },
  "securityDefinitions": {
    "JWT": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  }
}